requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
dj-database-url==2.2.0
Django==5.0.2
django-allauth==0.63.6
django-cors-headers==4.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flatbuffers==24.3.25
fonttools==4.51.0
humanfriendly==10.0
idna==3.8
imageio==2.35.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.3.7
lazy_loader==0.4
llvmlite==0.43.0
matplotlib==3.9.0
mpmath==1.3.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.19.0
opencv-python-headless==4.10.0.84
packaging==24.0
pandas==2.2.2
pillow==10.2.0
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==5.28.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyMatting==1.1.12
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sqlparse==0.4.4
sympy==1.13.2
threadpoolctl==3.5.0
tifffile==2024.8.30
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
